pyproject.toml
[tool.poetry]
name = "twittersentimentanalysis"
version = "0.1.0"
description = ""
authors = ["Varun Pius Rodriguus <wraith@test.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tweepy = "^4.13.0"
ruamel-yaml = "^0.17.21"
kafka-python = "^2.0.2"
pyspark = "^3.3.2"
findspark = "^2.0.1"
textblob = "^0.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
